Move get_patch_mbox() function to misc.py
We now have the misc.py file which should be the home of the
functions. Init file's purpose is to mark the module, not to
contain a bunch of code, and the file isn't even needed with
the newer versions of Python so let's get rid of its contents.

Fixes #291
