+++ skt/executable.py
@@ -36,7 +36,7 @@ import skt.reporter
 import skt.runner
 from skt.kernelbuilder import KernelBuilder, CommandTimeoutError, ParsingError
 from skt.kerneltree import KernelTree, PatchApplicationError
-from skt.misc import (join_with_slash, get_patch_name,
+from skt.misc import (join_with_slash, get_patch_name, get_patch_mbox,
                       SKT_SUCCESS, SKT_FAIL, SKT_ERROR)
 
 DEFAULTRC = "~/.sktrc"
@@ -234,7 +234,7 @@ def cmd_merge(cfg):
                 for patch in cfg.get('pw'):
                     save_state(cfg, {'patchwork_%02d' % idx: patch})
                     report_string += '  - %s,\n' % get_patch_name(
-                        skt.get_patch_mbox(patch)
+                        get_patch_mbox(patch)
                     )
                     report_string += '    grabbed from %s\n' % patch
